requirements.txt
streamlit
requests
openai
python-docx
fpdf
